LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioworklet-interface/audioworklet-audioparam.https-expected.txt


PASS # AUDIT TASK RUNNER STARTED.
PASS Executing "Initializing AudioWorklet and Context"
PASS Executing "Verifying AudioParam in AudioWorkletNode"
PASS Audit report
PASS > [Initializing AudioWorklet and Context]
PASS < [Initializing AudioWorklet and Context] All assertions passed. (total 0 assertions)
PASS > [Verifying AudioParam in AudioWorkletNode]
PASS   Default gain value of gainWorkletNode is equal to 0.7070000171661377.
PASS   Value of gainWorkletParam after setter = 0.1 is equal to 0.10000000149011612.
FAIL X The rendered buffer: Expected 0 for all values but found 28798 unexpected values:
	Index	Actual
	[1]	-0.00020833333837799728
	[2]	-0.00041666667675599456
	[3]	-0.0006249999860301614
	[4]	-0.0008333333535119891
	...and 28794 more errors. assert_true: expected true got false
FAIL < [Verifying AudioParam in AudioWorkletNode] 1 out of 3 assertions were failed. assert_true: expected true got false
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 2 tasks were failed. assert_true: expected true got false
